requirements.txt
click>=8.1.0
colorama>=0.4.6
gitpython>=3.1.40
pyyaml>=6.0
questionary>=2.0.0
rich>=13.7.0
PyGithub>=2.8.0
tqdm>=4.65.0
click>=8.1.0
colorama>=0.4.6
flask
gitpython>=3.1.40
google-generativeai
PyGithub>=2.8.0
pyyaml>=6.0
questionary>=2.0.0
rich>=13.7.0
tqdm>=4.65.0
click
colorama
flask
gitpython
google-generativeai
PyGithub
pyyaml
questionary
rich
tqdm
click>=8.1.0
colorama>=0.4.6
flask
gitpython>=3.1.40
google-generativeai
PyGithub>=2.8.0
pyyaml>=6.0
questionary>=2.0.0
rich>=13.7.0
tqdm>=4.65.0